feat: split publicodes, partial handling of specific sector eligibility, update all data.
