+++ packages/data/programs/act-pas-a-pas.yaml
@@ -3,13 +3,14 @@ promesse: >-
   Développez votre stratégie bas carbone et mettez en œuvre votre plan de
   transition
 description: >-
-  L’ADEME soutient financièrement les entreprises volontaires dans la mise en
-  œuvre d'un modèle économique compatible avec une trajectoire bas carbone. 
+  L’ADEME soutient financièrement les entreprises qui souhaitent questionner
+  leur modèle économique pour identifier un nouveau modèle compatible avec une
+  trajectoire bas carbone. 
 
   Vous serez accompagné avec la méthode ACT Pas à Pas (Assessing Low Carbon
   Transition), une méthode structurante, progressive et sur le long terme qui
   aide les entreprises à réduire significativement leurs émissions de gaz à
-  effet de serre. 
+  effet de serre.
 description longue: >-
   L'étude ACT Pas à Pas se déroule en 12 à 18 mois selon les entreprises. Les
   grandes étapes de l'étude sont les suivantes :
