+++ packages/data/programs/aides-au-reemploi-des-emballages.yaml
@@ -7,7 +7,7 @@ description: >-
 description longue: >-
   Les dépenses éligibles sont :
 
-  •  Les investissements d'équipements et outils  pour permettre le passage au
+  •  Les investissements d'équipements et outils pour permettre le passage au
   réemploi
 
   •  Les expérimentations : tests à petite échelle préalable à l'investissement.
@@ -24,7 +24,7 @@ montant du financement: Jusqu'à 80% du coût de l'étude ou 60% de l'investisse
 objectifs:
 - description: >-
     Vérifiez l’éligibilité de votre projet en parcourant la documentation de
-    l’ADEME.
+    l’ADEME ou en contactant un conseiller.
   liens:
   - lien: >-
       https://agirpourlatransition.ademe.fr/entreprises/sites/default/files/R%C3%A9emploi%20emballages%20et%20contenants%20-%20Conditions%20d%27%C3%A9ligibilit%C3%A9%20et%20de%20financement%20-%202024.pdf
@@ -36,10 +36,18 @@ objectifs:
     Si vous souhaitez financer un investissement, vous devez avoir réalisé en
     amont les études justifiant l'intérêt économique et écologique de votre
     investissement. Elles seront à joindre à votre dossier.
+- description: Créez un compte ADEME pour pouvoir déposer votre dossier.
+  liens:
+  - lien: https://agirpourlatransition.ademe.fr/user/register
+    texte: Créer mon compte ADEME
 - description: >-
-    Déposez votre demande d'aide : informations administratives, description du
-    projet, détails techniques et financiers, etc. A ce stade, votre projet ne
-    doit pas avoir commencé !
+    Déposez votre dossier : informations administratives, description du projet,
+    détails techniques et financiers, etc. À ce stade, votre projet ne doit pas
+    avoir commencé !
+  liens:
+  - lien: >-
+      https://agirpourlatransition.ademe.fr/entreprises/aides-financieres/2024/aides-reemploi-emballages-contenants
+    texte: Déposer mon dossier
 - description: >-
     L'ADEME étudie votre demande d'aide. Si votre demande est acceptée, recevez
     un contrat de financement précisant le montant d’aide maximum, les
