+++ packages/data/programs/booster-eco-energie-tertiaire.yaml
@@ -111,7 +111,6 @@ publicodes:
     une de ces conditions:
     - questionnaire . objectif prioritaire . est rénover mon bâtiment
     - questionnaire . objectif prioritaire . est ma performance énergétique
-    - questionnaire . objectif prioritaire . est mon impact environnemental
   entreprise . est dans une zone géographique éligible:
     une de ces conditions:
     - région = Auvergne-Rhône-Alpes
