+++ packages/data/programs/diagnostic-changement-climatique.yaml
@@ -22,7 +22,9 @@ objectifs:
     Identifier les impacts directs sur vos activités ainsi que des impacts
     indirects (financiers et sociaux) et leurs évolutions à court, moyen et long
     terme.
-- description: Mettre en évidence les menaces posées par le changement climatique et les
+- description: >-
+    Mettre en évidence les menaces posées par le changement climatique et les 
+    opportunités potentiellement pour vos activités, produits et services.
 - description: >-
     Disposer d’un plan d’adaptation pour les risques climatiques au sein de
     votre entreprise.
