+++ packages/data/programs/etude-chaufferie-biomasse.yaml
@@ -11,12 +11,12 @@ description longue: >-
   combustibles d’origine végétale permettant la production de chaleur ou
   d’électricité. Première source d’énergie renouvelable utilisée en France, le
   bois doit fortement contribuer aux objectifs climatiques et énergétiques du
-  pays.
+  pays. 
 
-  Types de prestations concernés par le fonds chaleur :
+  Types de prestations concernés par le fonds chaleur : 
 
   - Etude de faisabilité d’une chaufferie biomasse/Ingénierie des installations
-  de production utilisant la biomasse en combustion.
+  de production utilisant la biomasse en combustion. 
 
   - Assistance à la maîtrise d’ouvrage pour la mise en place d’une chaufferie
   biomasse
