+++ packages/data/programs/etude-friches-polluees.yaml
@@ -96,4 +96,3 @@ publicodes:
   entreprise . a un objectif ciblé:
     une de ces conditions:
     - questionnaire . objectif prioritaire . est rénover mon bâtiment
-    - questionnaire . objectif prioritaire . est mon impact environnemental
