+++ packages/data/programs/etude-reemploi-reutilisation-reparation.yaml
@@ -102,7 +102,6 @@ publicodes:
     - code NAF niveau 1 . est R
   entreprise . a un objectif ciblé:
     une de ces conditions:
-    - questionnaire . objectif prioritaire . est la gestion des déchets
     - questionnaire . objectif prioritaire . est l'écoconception
   entreprise . est dans une zone géographique éligible:
     une de ces conditions:
