+++ packages/data/programs/fond-vert-territoires-d-industrie.yaml
@@ -1,0 +1,139 @@
+titre: Fond Vert - Territoires d'Industrie
+promesse: >-
+  Financez votre projet industriel d'intérêt général en matière de transition
+  écologique 
+description: >-
+  Vous portez un projet de développement industriel présentant un intérêt
+  général manifeste en termes de revitalisation économique et de transition
+  écologique ?  Le programme "Territoires d’industrie" soutient les entreprises
+  porteuses de projets d’investissements industriels ambitieux et exemplaires
+  sur le
+
+  plan environnemental, socio-économique et territorial. 
+description longue: >-
+  Les projets éligibles sont des projets d’investissements industriels
+  exemplaires sur le plan environnemental qui contribuent à l’émergence, au
+  renforcement ou à la réindustrialisation de chaînes de valeur industrielles
+  clés pour la transition écologique.
+
+  Afin d’être éligibles, les projets doivent être suffisamment matures. 
+
+
+  Par exemple :
+
+  • Bioéconomie pour le développement de produits biosourcés (chanvre, lin,
+  laine, bois, paille, bioéconomie marine, chimie, etc.) ;
+
+  • Nouvelles mobilités durables (vélos, vélos-cargos, véhicules intermédiaires,
+  etc.) ;
+
+  • Souveraineté alimentaire et relocalisation des filières agroalimentaires
+  (unités de
+
+  transformation locale de production, par exemple en lien avec un projet
+  alimentaire
+
+  territorial, etc.) ;
+
+  • Relocalisation de biens de consommation courants (habillement, mobilier,
+  etc.) avec un
+
+  process significativement plus respectueux de l’environnement que les
+  standards ;
+
+  • Productions industrielles contribuant au recyclage de matériaux ou matières
+  premières, à l’économie circulaire et au réemploi, ou à la valorisation de
+  déchets et co-produits
+
+  (équipements de la transition énergétique, de la rénovation du bâti, etc.) ; 
+
+  • Projets qui s’inscrivent dans une stratégie de diversification pour des
+  territoires fragiles dont
+
+  l’économie est impactée par le changement climatique (notamment territoires de
+  montagne). 
+
+  • Projets d’investissements contribuant au développement des
+
+  compétences (école de production, centre de formation, plateau technique,
+  etc.) 
+
+
+  À titre exceptionnel, un soutien spécifique et renforcé peut être donné aux
+  projets qui contribuent à la redynamisation de territoires « rebond industriel
+  » (territoires confrontés à de fortes difficultés industrielles ou des
+  restructurations importantes de chaînes de valeur)
+début de validité: 12/01/2024
+fin de validité: 15/12/2024
+illustration: images/TEE_eolienne.png
+opérateur de contact: ADEME
+url: >-
+  https://www.demarches-simplifiees.fr/commencer/fonds-vert-3-territoires-industrie
+nature de l'aide: financement
+montant du financement: Variable selon le projet
+objectifs:
+- description: >-
+    Vérifiez votre éligibilité géographique. Cette aide s'adresse aux
+    entreprises situées dans le périmètre d’un "Territoire d’industrie".
+  liens:
+  - lien: >-
+      https://agence-cohesion-territoires.gouv.fr/sites/default/files/inline-images/perimetre_des_territoires_dindustrie_2023-2027_high_0.png
+    texte: Territoires d'industrie 2023-2027
+- description: Confirmez l'éligibilité de votre projet en prenant contact avec l'ADEME.
+- description: >-
+    Déposez votre demande d'aide : informations administratives, description du
+    projet, détails techniques et financiers, étude de faisabilité, etc. A ce
+    stade, votre investissement ne doit pas avoir été realisé !
+- description: >-
+    L'ADEME étudie votre demande d'aide. Si votre demande est acceptée, recevez
+    un contrat de financement précisant le montant d’aide maximum, les
+    conditions et les modalités de versement.
+- description: Réalisez vos investissements.
+conditions d'éligibilité:
+  taille de l'entreprise:
+  - Toutes tailles
+  - Éligible aux micro-entreprises
+  secteur géographique:
+  - France et territoires d'outre-mer
+  secteur d'activité:
+  - Industrie
+  nombre d'années d'activité:
+  - Éligible à toutes les entreprises
+  autres critères d'éligibilité:
+  - >-
+    Cette aide est destinée aux entreprises situées sur les territoires
+    d'industrie 2023-2027
+  - Les projets doivent présenter une assiette minimale de dépenses de 400 000 €
+  - >-
+    L'opération pour laquelle vous sollicitez une aide financière ne doit pas
+    avoir commencé ou ne doit pas avoir donné lieu à des engagements fermes
+    (sous quelque forme que ce soit : marché signé, commande signée, devis
+    accepté…)
+publicodes:
+  entreprise . est ciblée:
+    toutes ces conditions:
+    - est éligible
+    - est dans un secteur d'activité ciblé
+    - a un objectif ciblé
+  entreprise . est éligible:
+    applicable si:
+      toutes ces conditions:
+      - dispositif . début de validité <= date du jour
+      - date du jour <= dispositif . fin de validité
+    valeur: oui
+  entreprise . est dans un secteur d'activité ciblé:
+    une de ces conditions:
+    - code NAF niveau 1 . est B
+    - code NAF niveau 1 . est C
+    - code NAF niveau 1 . est D
+    - code NAF niveau 1 . est E
+  entreprise . a un objectif ciblé:
+    une de ces conditions:
+    - questionnaire . objectif prioritaire . est rénover mon bâtiment
+    - questionnaire . objectif prioritaire . est la mobilité durable
+    - questionnaire . objectif prioritaire . est la gestion des déchets
+    - questionnaire . objectif prioritaire . est diminuer ma consommation d'eau
+    - questionnaire . objectif prioritaire . est ma performance énergétique
+    - questionnaire . objectif prioritaire . est former ou recruter
+    - questionnaire . objectif prioritaire . est mon impact environnemental
+    - questionnaire . objectif prioritaire . est l'écoconception
