+++ packages/data/programs/fonds-tourisme-durable.yaml
@@ -34,7 +34,7 @@ objectifs:
   - lien: >-
       https://agirpourlatransition.ademe.fr/entreprises/sites/default/files/Fonds%20Tourisme%20Durable%20-%20Conditions%20d%27%C3%A9ligibilit%C3%A9%20et%20de%20financement%20-%202024.pdf
     texte: Conditions éligibilité et financement
-- description: 'Vérifiez l''éligibilité géographique de votre entreprise. '
+- description: Vérifiez l'éligibilité géographique de votre entreprise.
   liens:
   - lien: >-
       https://agirpourlatransition.ademe.fr/entreprises/sites/default/files/Fonds%20Tourisme%20Durable%20-%20P%C3%A9rim%C3%A8tre%20communes%20%C3%A9ligibles%20-%2001-06-2023.xlsx
