+++ packages/data/programs/formation-engager-entreprise-transition-ecologique.yaml
@@ -1,23 +1,12 @@
-titre: Formations ADEME
-promesse: Engager son entreprise dans la transition écologique
+titre: Formation ADEME
+promesse: Réalisez une initiation aux enjeux de la transition écologique
 description: >-
-  Plongez au cœur des enjeux écologiques et des scénarios de transition de
-  l'ADEME pour (re)découvrir leurs conséquences et leurs opportunités pour votre
-  entreprise
-description longue: >-
-  A la fin de la formation, les participants sauront/seront capables de :
-
-
-  - Avoir la conviction qu'il faut agir,
-
-  - Identifier les principales transformations de la société, de l'économie, des
-  territoires et les enjeux pour leur entreprise,
-
-  - Anticiper le changement, les risques et les opportunités,
-
-  - Identifier les premiers leviers d’action pour se mettre en mouvement.
-début de validité: 01/01/2024
-fin de validité: 31/12/2024
+  Suivez une formation en ligne pour initier la transition écologique de votre
+  entreprise. Vous découvrirez les principales transformations de la société, de
+  l'économie, des territoires et les enjeux de la transition écologique pour les
+  entreprises. Puis en seconde partie, vous serez invité à identifier les
+  risques et les opportunités pour votre entreprise ainsi que les premiers
+  leviers d’action pour se mettre en mouvement.
 illustration: images/TEE_eolienne.png
 opérateur de contact: ADEME
 url: >-
@@ -30,16 +19,20 @@ durée de l'accompagnement: |-
   2h15
 activable en autonomie: oui
 objectifs:
-- description: Créez votre compte sur la plateforme ADEME Formations.
+- description: Prenez connaissance du programme de la formation
+  liens:
+  - lien: >-
+      https://formations.ademe.fr/tmp/01001119-1720429428-9999999-9a7cecf0ea4d98c67fe5e83d4a1d40b9.pdf
+    texte: télécharger le programme
+- description: >-
+    Inscrivez-vous sur la plateforme ADEME Formations. Vous pouvez ajouter
+    plusieurs participants.
   liens:
   - lien: https://formations.ademe.fr/solutions/beforeregister.php?sessionid=7866
     texte: formations.ademe.fr
 - description: >-
-    Ajoutez un ou plusieurs participants, ainsi que les coordonnées de votre
-    entreprise.
-- description: Vous pourrez nous signaler tout besoin spécifique ou situation de handicap.
-- description: Recevez votre récapitulatif de votre formation et une validation.
-- description: Réalisez votre formation auprès des experts de l'ADEME
+    Obtenez le lien de la formation à réaliser en ligne en toute autonomie à
+    l'aide de vidéos et de quizz.
 conditions d'éligibilité:
   taille de l'entreprise:
   - Toutes tailles
@@ -50,20 +43,11 @@ conditions d'éligibilité:
   - Tous secteurs d’activité
   nombre d'années d'activité:
   - Éligible à toutes les entreprises
-  autres critères d'éligibilité:
-  - Aucun prérequis n'est exigé pour suivre cette formation
 publicodes:
   entreprise . est ciblée:
     toutes ces conditions:
-    - est éligible
     - est dans un secteur d'activité ciblé
     - a un objectif ciblé
-  entreprise . est éligible:
-    applicable si:
-      toutes ces conditions:
-      - dispositif . début de validité <= date du jour
-      - date du jour <= dispositif . fin de validité
-    valeur: oui
   entreprise . est dans un secteur d'activité ciblé:
     une de ces conditions:
     - code NAF niveau 1 . est A
@@ -90,4 +74,3 @@ publicodes:
   entreprise . a un objectif ciblé:
     une de ces conditions:
     - questionnaire . objectif prioritaire . est former ou recruter
-    - questionnaire . objectif prioritaire . est mon impact environnemental
