+++ packages/data/programs/formations-tee.yaml
@@ -7,7 +7,7 @@ description: >-
   animées par des experts de la transition écologique et énergétique. Ces
   formations abordent de nombreux sujets relatifs aux enjeux de la
   responsabilité sociétale des entreprises (RSE) : environnement, enjeux
-  sociaux, achats responsables…
+  sociaux, achats responsables… 
 description longue: >-
   Ce cursus Transition Écologique et Énergétique vous permettra de :
