+++ packages/data/programs/gestion-des-biodechets.yaml
@@ -16,7 +16,7 @@ description longue: >-
 
 
   En pratique, les biodéchets à traiter sont :
-
+   
   Déchets verts : tontes de pelouse et fauchage, feuilles mortes, tailles
   d’arbustes, haies et brindilles ou encore déchets ligneux issus de l’élagage
   et de l’abattage d’arbres et de haies.
