+++ packages/data/programs/investissement-contrat-patrimonial-enr-r.yaml
@@ -95,7 +95,6 @@ publicodes:
     - code NAF niveau 1 . est U
   entreprise . a un objectif ciblé:
     une de ces conditions:
-    - questionnaire . objectif prioritaire . est diminuer ma consommation d'eau
     - questionnaire . objectif prioritaire . est ma performance énergétique
   entreprise . est dans une zone géographique éligible:
     une de ces conditions:
