+++ packages/data/programs/investissement-contre-le-gaspillage-alimentaire.yaml
@@ -120,7 +120,6 @@ publicodes:
   entreprise . a un objectif ciblé:
     une de ces conditions:
     - questionnaire . objectif prioritaire . est la gestion des déchets
-    - questionnaire . objectif prioritaire . est l'écoconception
   entreprise . est dans une zone géographique éligible:
     une de ces conditions:
     - région = Bourgogne-Franche-Comté
