+++ packages/data/programs/investissement-gestion-des-biodechets.yaml
@@ -70,7 +70,6 @@ publicodes:
   entreprise . a un objectif ciblé:
     une de ces conditions:
     - questionnaire . objectif prioritaire . est la gestion des déchets
-    - questionnaire . objectif prioritaire . est l'écoconception
   entreprise . est dans une zone géographique éligible:
     une de ces conditions:
     - région = Auvergne-Rhône-Alpes
