+++ packages/data/programs/investissement-recharge-vehicules-electriques.yaml
@@ -87,5 +87,3 @@ publicodes:
   entreprise . a un objectif ciblé:
     une de ces conditions:
     - questionnaire . objectif prioritaire . est la mobilité durable
-    - questionnaire . objectif prioritaire . est ma performance énergétique
-    - questionnaire . objectif prioritaire . est l'écoconception
