+++ packages/data/programs/investissement-reemploi-reutilisation-reparation.yaml
@@ -97,7 +97,7 @@ publicodes:
     - code NAF niveau 1 . est S
   entreprise . a un objectif ciblé:
     une de ces conditions:
-    - questionnaire . objectif prioritaire . est la gestion des déchets
+    - questionnaire . objectif prioritaire . est l'écoconception
   entreprise . est dans une zone géographique éligible:
     une de ces conditions:
     - région = Auvergne-Rhône-Alpes
