+++ packages/data/programs/investissement-renouvellement-forestier.yaml
@@ -68,5 +68,4 @@ publicodes:
     - code NAF niveau 1 . est A
   entreprise . a un objectif ciblé:
     une de ces conditions:
-    - questionnaire . objectif prioritaire . est mon impact environnemental
-    - questionnaire . objectif prioritaire . est l'écoconception
+    - questionnaire . objectif prioritaire . est rénover mon bâtiment
