+++ packages/data/programs/investissement-systemes-solaires-combines.yaml
@@ -114,6 +114,4 @@ publicodes:
     - code NAF niveau 1 . est U
   entreprise . a un objectif ciblé:
     une de ces conditions:
-    - questionnaire . objectif prioritaire . est rénover mon bâtiment
-    - questionnaire . objectif prioritaire . est diminuer ma consommation d'eau
     - questionnaire . objectif prioritaire . est ma performance énergétique
