+++ packages/data/programs/ormat-2024.yaml
@@ -109,7 +109,6 @@ publicodes:
   entreprise . a un objectif ciblé:
     une de ces conditions:
     - questionnaire . objectif prioritaire . est la gestion des déchets
-    - questionnaire . objectif prioritaire . est l'écoconception
   entreprise . est dans une zone géographique éligible:
     une de ces conditions:
     - région = Auvergne-Rhône-Alpes
