+++ packages/data/programs/renovation-energetique.yaml
@@ -3,7 +3,7 @@ promesse: Bénéficiez d'un crédit d’impôt pour la rénovation énergétique
 description: >-
   Vous bénéficiez d’un crédit d’impôt pour la rénovation énergétique (travaux
   d’isolation, chauffe-eau solaire, pompe à chaleur, ventilation mécanique…) de
-  vos bâtiments à usage tertiaire.
+  vos bâtiments à usage tertiaire. 
 description longue: >-
   Ce crédit d'impôt pour les PME s'applique aux dépenses de rénovation, engagées
   entre le 1er janvier 2023 et le 31 décembre 2024, sur des bâtiments du
@@ -11,7 +11,7 @@ description longue: >-
   travaux ont débuté.
 
 
-  Les dépenses éligibles sont :
+  Les dépenses éligibles sont : 
 
 
   - Acquisition et pose d'un système d'isolation thermique en rampant de toiture
@@ -21,7 +21,7 @@ description longue: >-
   pignon, par l'extérieur ou l’intérieur,
 
   - Acquisition et pose d'un système d'isolation thermique en toiture terrasse
-  ou couverture de pente inférieure à 5 %,
+  ou couverture de pente inférieure à 5 %, 
 
   - Acquisition et pose d'un chauffe-eau solaire collectif ou d'un dispositif
   solaire collectif pour la production d'eau chaude sanitaire,
