+++ packages/data/programs/repar-acteurs.yaml
@@ -1,4 +1,4 @@
-titre: Repar'acteurs
+titre: Répar'acteurs
 promesse: >-
   Rejoignez un réseau de professionnels de la réparation, devenez artisan
   Répar’acteurs !
@@ -13,13 +13,13 @@ description longue: >-
   entreprises (RNE) sont éligibles à la marque répar’acteurs.
 
 
-  Pour être éligible, votre entreprise doit remplir l'une de ces conditions  :
+  Pour être éligible, votre entreprise doit remplir l'une de ces conditions  : 
 
-  • Détenir la qualité Artisan ou le titre de Maître Artisan,
+  • Détenir la qualité Artisan ou le titre de Maître Artisan, 
 
   • Bénéficier d'un diplôme ou un titre certifié dans l’activité de réparation,
 
-  • Exercer depuis 3 années d’expérience professionnelle.
+  • Exercer depuis 3 années d’expérience professionnelle. 
 
 
   Les activités éligibles à la marque répar’acteurs sont toutes celles
@@ -42,7 +42,7 @@ objectifs:
 - description: >-
     Pour obtenir le label Répar'acteurs, créez un compte, complétez une charte
     d'engagement et chargez vos pièces justificatives. Vous recevrez dans les
-    prochains jours la confirmation de votre  éligibilité.
+    prochains jours la confirmation de votre éligibilité.
   liens:
   - lien: >-
       https://www.demarches-simplifiees.fr/commencer/devenir-repar-acteurs-formulaire-d-engagement
